refactor: User 도메인의 Query 및 Command 분리
